Added routes modifed tables to handle timeStamps

+++ spring-boot-api-server/src/main/java/com/niantic/repositories/MySqlPostDao.java
@@ -9,8 +9,7 @@ import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Repository;
 
 import javax.sql.DataSource;
-import java.sql.PreparedStatement;
-import java.sql.Statement;
+import java.sql.*;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -43,9 +42,9 @@ public class MySqlPostDao implements PostDao {
             int userId = row.getInt("user_id");
             String imgUrl = row.getString("img_url");
             String title = row.getString("title");
-            String captions = row.getString("caption");
+            String captions = row.getString("captions");
             int reactions = row.getInt("reactions");
-            int albumId = row.getInt("album_id");
+            Integer albumId = (Integer) row.getObject("album_id");
             LocalDateTime createdAt = row.getTimestamp("created_at").toLocalDateTime();
 
             Post post = new Post(postId, userId, imgUrl, title, captions, reactions, albumId, createdAt);
@@ -75,9 +74,9 @@ public class MySqlPostDao implements PostDao {
             userId = row.getInt("user_id");
             String imgUrl = row.getString("img_url");
             String title = row.getString("title");
-            String captions = row.getString("caption");
+            String captions = row.getString("captions");
             int reactions = row.getInt("reactions");
-            int albumId = row.getInt("album_id");
+            Integer albumId = (Integer) row.getObject("album_id");
             LocalDateTime createdAt = row.getTimestamp("created_at").toLocalDateTime();
 
             Post post = new Post(postId, userId, imgUrl, title, captions, reactions, albumId, createdAt);
@@ -88,7 +87,7 @@ public class MySqlPostDao implements PostDao {
         return posts;
     }
 
-    public List<Post> getPostsByAlbumId(int albumId)
+    public List<Post> getPostsByAlbumId(Integer albumId)
     {
         List<Post> posts = new ArrayList<>();
 
@@ -106,9 +105,9 @@ public class MySqlPostDao implements PostDao {
             int userId = row.getInt("user_id");
             String imgUrl = row.getString("img_url");
             String title = row.getString("title");
-            String captions = row.getString("caption");
+            String captions = row.getString("captions");
             int reactions = row.getInt("reactions");
-            albumId = row.getInt("album_id");
+            albumId = (Integer) row.getObject("album_id");
             LocalDateTime createdAt = row.getTimestamp("created_at").toLocalDateTime();
 
             Post post = new Post(postId, userId, imgUrl, title, captions, reactions, albumId, createdAt);
@@ -137,9 +136,9 @@ public class MySqlPostDao implements PostDao {
             int userId = row.getInt("user_id");
             String imgUrl = row.getString("img_url");
             String title = row.getString("title");
-            String captions = row.getString("caption");
+            String captions = row.getString("captions");
             int reactions = row.getInt("reactions");
-            int albumId = row.getInt("album_id");
+            Integer albumId = (Integer) row.getObject("album_id");
             LocalDateTime createdAt = row.getTimestamp("created_at").toLocalDateTime();
 
             post = new Post(postId, userId, imgUrl, title, captions, reactions, albumId, createdAt);
@@ -153,22 +152,27 @@ public class MySqlPostDao implements PostDao {
     {
         String sql = """
                     INSERT INTO posts
-                    (user_id, img_url, title, captions, reactions, album_id, created_at)
-                    VALUES (?,?,?,?,?,?,?)
+                    (user_id, img_url, title, captions, reactions, album_id)
+                    VALUES (?,?,?,?,DEFAULT,?)
                     """;
 
         KeyHolder keyHolder = new GeneratedKeyHolder();
 
         jdbcTemplate.update(connection -> {
+
             PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
 
             statement.setInt(1, post.getUserId());
             statement.setString(2, post.getImgUrl());
             statement.setString(3, post.getTitle());
             statement.setString(4, post.getCaptions());
-            statement.setInt(5, post.getReactions());
-            statement.setInt(6, post.getAlbumId());
-//            statement.setDate(7, (java.sql.Date) post.getCreatedAt());
+
+            if(post.getAlbumId() == null)
+            {
+                statement.setNull(5, Types.INTEGER);
+            } else {
+                statement.setInt(5, post.getAlbumId());
+            }
 
             return statement;
         }, keyHolder);
@@ -178,37 +182,58 @@ public class MySqlPostDao implements PostDao {
         return getPostById(newId);
     }
 
-    public void updatePost(int postId, Post post)
+    public boolean updatePost(int postId, Post post)
     {
-        String sql = """
-                UPDATE Posts
-                SET
-                    user_id = ?,
-                    img_url = ?,
-                    title = ?,
-                    captions = ?,
-                    reactions = ?,
-                    album_id = ?,
-                WHERE
-                    post_id = ?;
-                """;
+        try
+        {
+            List<Object> sqlColumns = new ArrayList<>();
+            StringBuilder sql = new StringBuilder("Update posts SET ");
 
-        jdbcTemplate.update(sql
-                , post.getUserId()
-                , post.getImgUrl()
-                , post.getTitle()
-                , post.getCaptions()
-                , post.getReactions()
-                , post.getAlbumId()
-                , postId);
+            if (post.getImgUrl() != null) {
+                sql.append("img_url = ?, ");
+                sqlColumns.add(post.getImgUrl());
+            }
+            if (post.getTitle() != null) {
+                sql.append("title = ?, ");
+                sqlColumns.add(post.getTitle());
+            }
+            if (post.getCaptions() != null) {
+                sql.append("captions = ?, ");
+                sqlColumns.add(post.getCaptions());
+            }
+            if (post.getAlbumId() != null) {
+                sql.append("album_id = ?, ");
+                sqlColumns.add(post.getAlbumId());
+            }
+
+            sql.setLength(sql.length() - 2);
+            sql.append(" WHERE post_id = ?;");
+            sqlColumns.add(postId);
+
+            jdbcTemplate.update(sql.toString(), sqlColumns.toArray());
+            return true;
+
+        }
+        catch(Exception e)
+        {
+            return false;
+        }
     }
 
-    public void deletePost(int postId)
+    public boolean deletePost(int postId)
     {
-        String sql = """
+        var postToDelete = getPostById(postId);
+
+        if(postToDelete != null)
+        {
+            String sql = """
                 DELETE FROM posts
                 WHERE post_id = ?
                 """;
-        jdbcTemplate.update(sql, postId);
+            jdbcTemplate.update(sql, postId);
+
+            return true;
+        }
+        return false;
     }
 }
